data_observability_agent/backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Agent framework (1.x family — all must be upgraded together)
langgraph==1.1.10
langchain==1.2.17
langchain-openai==1.2.1
langchain-core==1.3.3
langchain-mcp-adapters==0.2.2

# MCP + Airflow (mcp>=1.9.2 pulled by langchain-mcp-adapters)
apache-airflow-client==3.0.0

# LLM / embeddings / reranking
openai==2.36.0          # 2.x required by langchain-openai 1.x
cohere==5.21.1          # pinned to 5.x — AsyncClientV2 API used in retriever

# Database
asyncpg==0.29.0
pgvector==0.3.5

# Guardrails + PII
presidio-analyzer==2.2.362
presidio-anonymizer==2.2.362
spacy==3.7.6
# After install: python -m spacy download en_core_web_sm

# Observability
langsmith==0.8.3

# Scheduling
apscheduler==3.10.4

# Config + utils
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
pyyaml==6.0.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
